Use full bounds not just line number
