+++ editor/src/components/custom-code/code-file.ts
@@ -53,6 +53,7 @@ import { getProjectFileByFilePath } from '../assets'
 import type { EditorDispatch } from '../editor/action-types'
 import { StylingOptions } from 'utopia-api'
 import type { Emphasis, Focus, Icon, Styling } from 'utopia-api'
+import type { Bounds } from 'utopia-vscode-common'
 
 type ModuleExportTypes = { [name: string]: ExportType }
 
@@ -227,17 +228,17 @@ export function isDefaultComponentDescriptor(
 export interface ComponentDescriptorFromDescriptorFile {
   type: 'DESCRIPTOR_FILE'
   sourceDescriptorFile: string
-  lineNumber: number | null
+  bounds: Bounds | null
 }
 
 export function componentDescriptorFromDescriptorFile(
   sourceDescriptorFile: string,
-  lineNumber: number | null,
+  bounds: Bounds | null,
 ): ComponentDescriptorFromDescriptorFile {
   return {
     type: 'DESCRIPTOR_FILE',
     sourceDescriptorFile: sourceDescriptorFile,
-    lineNumber: lineNumber,
+    bounds: bounds,
   }
 }
