+++ editor/src/core/vscode/vscode-bridge.ts
@@ -6,7 +6,7 @@ import type {
   SelectedElementChanged,
   UpdateDecorationsMessage,
   ForceNavigation,
-  BoundsInFile,
+  Bounds,
 } from 'utopia-vscode-common'
 import {
   boundsInFile,
@@ -178,7 +178,7 @@ export function sendMessage(message: FromUtopiaToVSCodeMessage) {
   vscodeIFrame?.postMessage(message, { targetOrigin: '*' })
 }
 
-export function sendOpenFileMessage(filePath: string, bounds?: BoundsInFile) {
+export function sendOpenFileMessage(filePath: string, bounds: Bounds | null) {
   sendMessage(toVSCodeExtensionMessage(openFileMessage(filePath, bounds)))
 }
